# Optimize casting empty strings