Summary:
With this diff when casting empty string with in a try or try_cast
we avoid the throwing exceptions by pre-checking empty strings.
1.8s ->   655.78ms
not that it is still much higher than valid hence we need to investigate more the
source of slowdown.
```
============================================================================
[...]hmarks/ExpressionBenchmarkBuilder.cpp     relative  time/iter   iters/s
============================================================================
cast_int##try_invalid_empty_input                         655.78ms      1.52
cast_int##try_invalid_nan                                    1.61s   621.92m
cast_int##try_valid                                         2.10ms    475.22
cast_int##valid                                             2.19ms    457.07
----------------------------------------------------------------------------
```

Differential Revision: D47348604